Check pre-commit is already installed or not

+++ scripts/install_commit_hooks.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env sh
 
-if [ -d '.git/hooks' ]; then
+if [ -d '.git/hooks' ] && [ ! -f '.git/hooks/pre-commit' ] ; then
   ln -s ../../git-hooks/pre-commit .git/hooks
+  echo 'git pre-commit hook is installed'
 else
-  echo "Cannot find .git/hooks"
-  exit 1
+  echo 'git pre-commit hook is already installed'
 fi
